refactor: 폴더 구조 단순화, 컨벤션에 맞게 수정
