+++ src/marketing/HowSection.stories.tsx
